Build .pm from TT2 template IFF inc/.author exists
This will allow creation of lib/Net/Twitter/Lite.pm from TT2 template in
Makefile.PL for users who grab the project from the git repository without
requiring the CPAN distribution to have a TT2 requiremnt for installation.
(The resulting .pm file will be included in the CPAN distribution, even though
it is not included in the git repository.